Create program_to_find_index_of_an_array_element
ava program to find the index of an array element in an array of size 10. The program should not use any function other than main ( ) functions.
Definition of Done:
DoD 1: The program should ask the user to enter the elements of the array.
DoD 2: The program should ask the user to enter a number to search.
DoD 3: The program should display the elements of the array entered.
DoD 4: The program should display the index of the number if the item is present or display -1 of the elements is not present.