Create program_to_find_distance_between_two_points
a program to calculate the distance between two points (x1, y1) and (x2,
y2). All numbers and return values should be of type double.
Definition of Done:
DoD 1: Two java files to be defined. One for class definitions and another for the
application

FOCP II Lab Manual (CSL 108)
2022-2023

DoD 2: A class point is defined with two float variables for x1 and x2 and the following
functionality:
i. Non-parametrized and parameterized constructors are defined.
ii. Get and set methods are defined for all the instance variables.
iii. Distance function is defined to calculate the distance between two points.
iv. Display function is defined with width of 7 and precision of 2.
Write this program with a static method definition for calculating the distance between
two points.