Create exception_handling_program_3
Create a three-level hierarchy of exceptions. Now create a base-class A with a method that throws an exception at the base of your hierarchy. Inherit B from A and override the method so it throws an exception at level two of your hierarchy. Repeat by inheriting class C from B. In main( ), create a C and upcast it to A, then call the method.
      Definition of Done:
1.  	Create three classes A,B,C
2.  	Display appropriate messages according to the exceptions.
3.  	The program needs to be well commented